Added initial try of generating the DZI
This basically uses Deep Zoom Composer and massages its project file a
bit. For the first try that was easier than trying to generate the tile
pyramid by hand. The downside is that the generated tiles have one layer
too many which is scaled to twice the size, for whichever reason. In the
public version that was removed and will never be requested either. But
it also means that the 100 % view is scaled down from a scaled-up
version, which makes it slightly blurrier than the original.